Refactor code out of models and into service layer
